requirements.txt
django
django_extensions
selenium
requests
djangorestframework